Switch open mp variable to integer.

+++ src/openmp_example/OpenMP_example.cpp
@@ -17,7 +17,9 @@ std::vector<double> OpenMP_example::add(const std::vector<double>& a, const std:
 {
     spdlog::info("Max OpenMP threads: {0}", getMaxThreads());
 
-    unsigned long long i {0};
+
+    int i {0}; // As msvc only supports openMP 2.0 so far, this variable needs to be an integer.
+
     std::vector<double> result (a.size());
 #if defined(_OPENMP)
 #pragma omp parallel for
